requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
tensorflow-cpu==2.12.0   # 또는 Cloud에서 설치 가능한 버전
mtcnn
pillow
pandas
